updated numerous pages to prepare CartContext. Also switched file name of CartContext.js to CartContext.jsx.
