+++ IT-Wizards-UI/src/App.jsx
@@ -18,21 +18,26 @@ import CartProvider from './components/CartContext';
 const App = () => {
   const router = createBrowserRouter(
     createRoutesFromElements(
-      <CartProvider>
-        <Route path="/" element={<MainLayout />}>
-          <Route index element={<HomePage />} />
-          <Route path="/items" element={<ItemsPage />} />
-          <Route path="/addItems" element={<AddItemForm />} />
-          <Route path="/items/editItem/:id" element={<EditItemForm />} />
-          <Route path="/itemCategories" element={<ItemCategoriesPage />} />
-          <Route path="/addItemCategories" element={<AddItemCategoryForm />} />
-          <Route path="/itemCategories/:id"element={<EditItemCategoriesForm />}/>
-          <Route path="/items/:id" element={<ItemDetailsPage />} />
-        </Route>
-      </CartProvider>
+      <Route path="/" element={<MainLayout />}>
+        <Route index element={<HomePage />} />
+        <Route path="/items" element={<ItemsPage />} />
+        <Route path="/addItems" element={<AddItemForm />} />
+        <Route path="/items/editItem/:id" element={<EditItemForm />} />
+        <Route path="/itemCategories" element={<ItemCategoriesPage />} />
+        <Route path="/addItemCategories" element={<AddItemCategoryForm />} />
+        <Route
+          path="/itemCategories/:id"
+          element={<EditItemCategoriesForm />}
+        />
+        <Route path="/items/:id" element={<ItemDetailsPage />} />
+      </Route>
     )
   );
-  return <RouterProvider router={router} />;
+  return (
+    <CartProvider>
+      <RouterProvider router={router} />
+    </CartProvider>
+  );
 };
 
 export default App;
